greatly imporve memory usage of a game instance (5704 -> 788 bytes)

+++ lib/test.h
@@ -5,7 +5,7 @@ typedef enum {
   TSPIN_NO_LINES
 } test;
 
-u8 tspin_board[20][10] = {
+int tspin_board[20][10] = {
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
@@ -28,7 +28,7 @@ u8 tspin_board[20][10] = {
   { 1, 1, 1, 1, 1, 1, 1, 1, 1, 0 }
 };
 
-u8 tspin_no_lines_board[20][10] = {
+int tspin_no_lines_board[20][10] = {
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
   { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
@@ -51,6 +51,7 @@ u8 tspin_no_lines_board[20][10] = {
   { 1, 1, 1, 1, 1, 1, 1, 1, 1, 0 }
 };
 
+/*
 void apply_test_board(cetris_game* g, test t) {
   u8 (*board)[20][10];
   switch (t) {
@@ -73,3 +74,4 @@ void apply_test_board(cetris_game* g, test t) {
     }
   }
 }
+*/
